Respect the order in which -b and -w are given
With this, only -w specified after -b will include window borders.
This enables the user to include borders for only some of the windows.

This can also be a double-edged sword, in that now the user has finer
control, but (s)he must be fully aware of the effects of the order in
which options are given.  Yes, you should.
However, the user will most likely want to specify -b at the very
first, which gives the old behavior of a "global" switch.