Support arbitrary command line via format string
This really is the way to be free, for everyone.  Thanks to the simple
command line partition mechanism, this is implemented as a natural
extension to the original syntax, in a completely backward compatible
manner.  This new syntax in fact supersedes the old syntax, but the old
syntax is more convenient, and is for now preserved.

Also worth mentioning, is that the cast command line is now executed via
`exec', since there is no command name verification for the format string
syntax.  Without `exec', internal functions would be accessible as the
command name, which is undesirable.