Change the option -m to require an integer arg
Always trimming the region is wrong.  I'd rather not trim it at all, but
the default ffmpeg command encodes to x264 which requires dimensions to
be divisible by 2.  For this reason, we keep the default behavior
unchanged, but allow the user to change or disable trimming.

This change introduces backward incompatibility, but the failure due to
this change will be noisy, and it's easy enough for the user to change
the old `-m' to `-m 16'.

This commit also fixes a bug the we did not adjust the right and bottom
offsets when the dimensions had changed.  This geometry post-processing
is still pretty lame nevertheless.