Support selecting specific XINERAMA heads
This should work for multi-head Xinerama setups.  The idea is simple,
but the implementation may seem a little more complicated than usual,
due to the limitations of bash as a language.  Hopefully, this is not
too ugly.
This is also the first time `grep' and `sed' are used in this script,
because the 660+ lines of `xdpyinfo` output makes the pattern matching
of bash a bit slow when using globbing, *very* slow when using regexp.