Refactor command line syntax
There is no more "double hyphen syntax". And there is rarely any need to
put "%" before the command. Enter sub-commands.

Sub-commands are supposed to be brief and to the point. Unlike before,
the name of the sub-command does not have to be the same as the actual
command executed by it. Each sub-command is associated with a short
"docstring", which will be shown when `-l` is used. Each sub-command can
do anything with the remaining command line arguments.

The "format string syntax", in this light, is simply yet another
sub-command, namely '%'.

To specify geospec, one must now use '-g <geospec>', which can be
repeated as any other option. This makes it clear and easy to figure out
where the sub-command starts.

Note this is a backward-compatibility breaking change.

Also, remove much of README. It's pretty useless/outdated and is a
burden to maintenance.