New option -#: select window by ID
This was at first motivated by the idea of capturing a moving window.
The idea is, since the window can be identified with an ID, we can
repeated get its geometry and take a screenshot of it.

So this has been implemented. The idea though doesn't really work. Try,

    for ((i=0; i<30 ;i++)); do
        ffcast -# 0x1200001 png /tmp/$i.png;
    sleep 0.1; done

Since there is a delay between the time we get the window geometry and
the time ffmpeg acutaly takes a screenshot of that region, a fast-moving
window will have moved in that time frame and it's reflected in the
screenshot. The serendipity here though, is that this can be used to
easily capture the "flickering" at the edge of a moving window.