Remove options -m -p; refactor sub-command code
Remove -p because it's now what ffcast does by default when no command
is specified. The default command before, which was a ffmpeg command
line for screen recording, is now made into a sub-command 'rec'.

Remove -m because trimming is rarely needed. The reason for its
introduction was to make ffmpeg happy. Now we make the new sub-command
'rec' accept an optional -m option, and use the ffmpeg "crop" video
filter to do this thing.

The sub-command execution is made into a function run_subcmd_or_print(),
mainly so it could be reused in a sub-command function.

A new convention is determined that our "official" sub-command functions
should be in the form of subcmd_*, but the more useful new convention is
to include a "docstring" in the form of ": 'usage: cmd ...'", so one can
quickly look it up in the output of `help cmd`.